pyproject.toml
[project]
name = "ai-agent-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "streamlit>=1.28.0",
    "langchain>=1.2.3",
    "langchain-ollama>=0.1.0",
    "langchain-openai>=0.2.0",
    "langgraph>=1.0.6",
    "openai>=1.0.0",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.0.0",
    "langchain-chroma>=1.1.0",
    "langchainhub>=0.1.21",
    "ipython>=9.9.0",
    "officeparser>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "python-pptx>=1.0.2",
    "pdf2image>=1.17.0",
    "langchain-huggingface>=1.2.0",
    "sentence-transformers>=5.2.0",
    "duckduckgo-search>=8.1.1",
    "ddgs>=9.10.0",
]

[tool.basedpyright]
venvPath = "."
venv = ".venv"
